[FIX] - Change model of the data binding to AddBuildingRoomVM
This change is important, beacuse with the old model, there was
a side case where the code woulndt save the room into the db
because it couldnt know if the popup was being invoked from
AddPopupVM (MainPopup) or it was being called on the AddBuildingVM.

So, in order to solve this issue, I created a new model that is
pretty much the same as the old one, but I added a new boolean property
that gives us the hint needed to know its origin path.

There is a lot of modified files because I had to change small
pieces of code from here and there. But all is related to this
model change.